requirements.txt
bcrypt==4.0.1
cffi==1.15.1
cryptography==39.0.1
future==0.18.3
iso8601==1.1.0
paramiko==3.0.0
pycparser==2.21
PyNaCl==1.5.0
pyserial==3.5
PyYAML==6.0
